updated configs to work with JSX
