+++ package.json
@@ -6,7 +6,8 @@
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
 		"start": "node /src/gachi.js",
-		"build": "babel src -d dist"
+		"build": "babel src -d dist",
+		"examples-jsx": "babel examples --relative -d ./"
 	},
 	"repository": {
 		"type": "git",
@@ -24,15 +25,13 @@
 	"homepage": "https://github.com/Zewasik/Gachi.js#readme",
 	"type": "module",
 	"dependencies": {
-		"@babel/plugin-transform-react-jsx": "^7.22.5",
 		"express": "^4.18.2",
 		"node": "18.16"
 	},
 	"devDependencies": {
 		"@babel/cli": "^7.22.9",
 		"@babel/core": "^7.22.9",
-		"@babel/preset-env": "^7.22.9",
-		"@babel/preset-react": "^7.22.5",
+		"@babel/plugin-transform-react-jsx": "^7.22.5",
 		"typescript": "^5.1.6"
 	}
 }
